+++ src/extension.ts
@@ -1,83 +1,57 @@
 import * as vscode from 'vscode';
-import * as path from 'path';
-
-import { findCsprojFile, findNamespace, toClassName, normalizeNamespace } from './core';
+import * as core from './core'
 
 export function activate(context: vscode.ExtensionContext) {
+	
+	// get project configurations
+	const config = vscode.workspace.getConfiguration("sharpSnippets");
+	
 	const completionProvider = vscode.languages.registerCompletionItemProvider({ scheme: 'file', language: 'csharp' }, {
 		async provideCompletionItems(document, position, token, context) {
-			const namespace : string | null = await getNamespace(document);
-			if(!namespace) {
+
+			const preferFileScopedNamespaceConfig = config.get<boolean>('preferFileScopedNamespaceConfig') ?? false;
+
+			// check if file is not a script .csx
+			if(!core.isCSharpFile(document)) {
 				return;
 			}
-			const className = getClassName(document);
-			var snippets = createSnippetsCompletions(namespace, className);
+			
+			// get csproj file path
+			const csproj = core.findCsprojFile(document)
+			if(!csproj) {
+				// no csproj file found, show error message and return
+				vscode.window.showErrorMessage('No csproj file found')
+				return;
+			}
 
-			return snippets;
+			// get namespace from csproj file
+			const namespace : string = await core.findNamespace(csproj, document)
+
+			// get class name from file name
+			const className = core.getClassName(document)
+
+			// check if file-scoped namespace is supported and user wants to use it
+			const useFileScopedNamespace = await core.isFileScopedNamespaceSupported(csproj) &&
+												preferFileScopedNamespaceConfig
+
+			// generate snippet text
+			const classCompletionText = core.generateClassSnippetText(namespace, className, 'class', useFileScopedNamespace)
+			const classCompletionItem = core.generateCompletionItem('class', classCompletionText, 'Generate class with namespace', vscode.CompletionItemKind.Snippet, true)
+			
+			const interfaceCompletionText = core.generateClassSnippetText(namespace, className, 'interface', useFileScopedNamespace)
+			const interfaceCompletionItem = core.generateCompletionItem('interface', interfaceCompletionText, 'Generate interface with namespace', vscode.CompletionItemKind.Snippet, true)
+			
+			const enumCompletionText = core.generateClassSnippetText(namespace, className, 'enum', useFileScopedNamespace)
+			const enumCompletionItem = core.generateCompletionItem('enum', enumCompletionText, 'Generate enum with namespace', vscode.CompletionItemKind.Snippet, true)
+			
+			const structompletionText = core.generateClassSnippetText(namespace, className, 'struct', useFileScopedNamespace)
+			const structompletionItem = core.generateCompletionItem('struct', structompletionText, 'Generate enum with namespace', vscode.CompletionItemKind.Snippet, true)
+			
+			return [classCompletionItem, interfaceCompletionItem, enumCompletionItem, structompletionItem]
 		}
 	});
 	context.subscriptions.push(completionProvider);
 }
 
-async function getNamespace(document: vscode.TextDocument) {
-	const csproj = findCsprojFile(document)
-	const namespace : string | null = await findNamespace(csproj)
-	if(!namespace) {
-		return null;
-	}
-	const projectRootRelativePath = path.relative(vscode.workspace.getWorkspaceFolder(document.uri)?.uri.path?? '', path.dirname(document.fileName))
-	return normalizeNamespace(namespace, projectRootRelativePath) ?? null
-}
-
-function getClassName(document: vscode.TextDocument) {
-	const fileName = path.basename(document.fileName, path.extname(document.fileName))
-	const className = toClassName(fileName)
-	return className
-}
-
-
-function createSnippetsCompletions(namespace: string, className: String) {
-	const classCompletion = new vscode.CompletionItem('class', vscode.CompletionItemKind.Snippet);
-	classCompletion.insertText = new vscode.SnippetString(
-	`namespace ${namespace};\n\n` +
-	`public class ${className}\n` +
-	`{\n` +
-	`\t$0\n` +
-	`}`
-	);
-	classCompletion.documentation = new vscode.MarkdownString('Creates a new class with namespace');
-	classCompletion.preselect = true;
-
-
-	const interfaceCompletion = new vscode.CompletionItem('interface', vscode.CompletionItemKind.Snippet);
-	interfaceCompletion.insertText = new vscode.SnippetString(
-	`namespace ${namespace};\n\n` +
-	`public interface ${className}\n` +
-	`{\n` +
-	`\t$0\n` +
-	`}`
-	);
-	interfaceCompletion.documentation = new vscode.MarkdownString('Creates a new interface with namespace');
-	interfaceCompletion.preselect = true;
-
-	const enumCompletion = new vscode.CompletionItem('enum', vscode.CompletionItemKind.Snippet);
-	enumCompletion.insertText = new vscode.SnippetString(
-	`namespace ${namespace};\n\n` +
-	`public enum ${className}\n` +
-	`{\n` +
-	`\t$0\n` +
-	`}`
-	);
-	enumCompletion.documentation = new vscode.MarkdownString('Creates a new enum with namespace');
-	enumCompletion.preselect = true;
-
-
-	return [classCompletion, interfaceCompletion, enumCompletion];
-}
-
-// function createModulesCompletions(namespace: string) {
-// 	const namespaceCompletion = new vscode.CompletionItem('namespace', vscode.CompletionItemKind.Module);
-// 	namespaceCompletion.insertText = namespace;
-// }
 
 export function deactivate() { }