+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+
+## [1.1.0]
+
+- New snippet for struct
+- Add ability to enable or disable file-scoped namespace from settings
+- Use file-scoped namespace just if csharp language version is supported by checking csproj target framework and lang version
+- Rewrite core functions and add some documents
+
 ## [1.0.0]
 
 - Initial release